Добавлены новые компонеты уведомление(NexusNotification) и диалоговое окно(NexusDialog). Данные компоненты добавлены в главный лейоут, добавлены сервисы и добавлены в методы на страницах

+++ SibCCSPETest.Admin/Page/GroupPage.razor.cs
@@ -9,6 +9,10 @@ namespace SibCCSPETest.Admin.Page
     {
         [Inject]
         public IAPIService? ServiceAPI { get; set; }
+        [Inject]
+        public INexusNotificationService? NotificationService { get; set; }
+        [Inject]
+        public INexusDialogService? DialogService { get; set; }
 
         private NexusTableGrid<GroupDTO>? NexusTable;
         private NexusTableGridEditMode EditMode = NexusTableGridEditMode.Single;
@@ -87,6 +91,7 @@ namespace SibCCSPETest.Admin.Page
             {
                 NexusTable!.Data.Add(data);
                 await NexusTable.SelectRow(data);
+                NotificationService.ShowSuccess("Группа добавлена", "Успех");
             }
         }
 
@@ -99,7 +104,8 @@ namespace SibCCSPETest.Admin.Page
                 if (index >= 0)
                     NexusTable.Data[index] = data;
                 await NexusTable.SelectRow(data);
-                await NexusTable.CancelEditRow(data);                
+                await NexusTable.CancelEditRow(data);
+                NotificationService.ShowSuccess("Группа изменена", "Успех");
             }
         }
 
@@ -108,8 +114,14 @@ namespace SibCCSPETest.Admin.Page
             if (NexusTable != null && NexusTable.SelectedRows.Count != 0)
             {
                 var data = NexusTable.SelectedRows.First();
-                await ServiceAPI!.GroupService.DeleteGroup(data.Id);
-                NexusTable.RemoveRow(data);
+                var settings = new NexusDialogSetting("Удаление группы", $"Вы уверены, что хотите удалить \"{data.Title}\" группу?", "Отменить", "Удалить");
+                var result = await DialogService.Show(settings);
+                if (result?.Canceled == false)
+                {
+                    await ServiceAPI!.GroupService.DeleteGroup(data.Id);
+                    NexusTable.RemoveRow(data);
+                    NotificationService.ShowSuccess("Группа удалена", "Успех");
+                }
             }
         }
