+++ SibCCSPETest.Admin/Page/QuestionPage.razor.cs
@@ -9,6 +9,10 @@ namespace SibCCSPETest.Admin.Page
     {
         [Inject]
         public IAPIService? ServiceAPI { get; set; }
+        [Inject]
+        public INexusNotificationService? NotificationService { get; set; }
+        [Inject]
+        public INexusDialogService? DialogService { get; set; }
 
         private NexusTableGrid<QuestionDTO>? NexusTable;
         private NexusTableGridEditMode EditMode = NexusTableGridEditMode.Single;
@@ -82,6 +86,7 @@ namespace SibCCSPETest.Admin.Page
                 NexusTable!.Data.Add(data);
                 await NexusTable.SelectRow(data);
                 await NexusTable.OnExpandRow(data);
+                NotificationService.ShowSuccess("Вопрос добавлен", "Успех");
             }
         }
 
@@ -95,6 +100,7 @@ namespace SibCCSPETest.Admin.Page
                     NexusTable.Data[index] = data;
                 await NexusTable.SelectRow(data);
                 await NexusTable.CancelEditRow(data);
+                NotificationService.ShowSuccess("Вопрос изменен", "Успех");
             }
         }
 
@@ -103,8 +109,14 @@ namespace SibCCSPETest.Admin.Page
             if (NexusTable != null && NexusTable.SelectedRows.Count != 0)
             {
                 var data = NexusTable.SelectedRows.First();
-                await ServiceAPI!.QuestionService.DeleteQuestion(data.Id);
-                NexusTable.RemoveRow(data);
+                var settings = new NexusDialogSetting("Удаление вопроса", $"Вопрос удалится вместе с ответами. Вы уверены, что хотите удалить \"{data.Title}\" вопрос?", "Отменить", "Удалить");
+                var result = await DialogService.Show(settings);
+                if (result?.Canceled == false)
+                {
+                    await ServiceAPI!.QuestionService.DeleteQuestion(data.Id);
+                    NexusTable.RemoveRow(data);
+                    NotificationService.ShowSuccess("Вопрос удален", "Успех");
+                }
             }
         }
