+++ SibCCSPETest.Admin/Page/SpecializationPage.razor.cs
@@ -9,6 +9,10 @@ namespace SibCCSPETest.Admin.Page
     {
         [Inject]
         public IAPIService? ServiceAPI { get; set; }
+        [Inject]
+        public INexusNotificationService? NotificationService { get; set; }
+        [Inject]
+        public INexusDialogService? DialogService { get; set; }
 
         private NexusTableGrid<SpecializationDTO>? NexusTable;
         private NexusTableGridEditMode EditMode = NexusTableGridEditMode.Single;
@@ -81,6 +85,7 @@ namespace SibCCSPETest.Admin.Page
             {
                 NexusTable!.Data.Add(data);
                 await NexusTable.SelectRow(data);
+                NotificationService.ShowSuccess("Специализация добавлена", "Успех");
             }
         }
 
@@ -94,6 +99,7 @@ namespace SibCCSPETest.Admin.Page
                     NexusTable.Data[index] = data;
                 await NexusTable.SelectRow(data);
                 await NexusTable.CancelEditRow(data);
+                NotificationService.ShowSuccess("Специализация изменена", "Успех");
             }
         }
 
@@ -102,8 +108,18 @@ namespace SibCCSPETest.Admin.Page
             if (NexusTable != null && NexusTable.SelectedRows.Count != 0)
             {
                 var data = NexusTable.SelectedRows.First();
-                await ServiceAPI!.SpecializationService.DeleteSpecialization(data.Id);
-                NexusTable.RemoveRow(data);
+                var settings = new NexusDialogSetting("Удаление специализации", $"Вы уверены, что хотите удалить \"{data.Title}\" специализацию?", "Отменить", "Удалить");
+                var result = await DialogService.Show(settings);
+                if (result?.Canceled == false)
+                {
+                    var isDeleted = await ServiceAPI!.SpecializationService.DeleteSpecialization(data.Id);
+                    if (isDeleted)
+                    {
+                        NexusTable.RemoveRow(data);
+                        NotificationService.ShowSuccess("Специализация удалена", "Успех");
+                    }
+                    else NotificationService.ShowError("Удалить специализацию нельзя из-за наличия связей", "Ошибка");
+                }
             }
         }
 
@@ -111,6 +127,46 @@ namespace SibCCSPETest.Admin.Page
         {
             var data = NexusTable!.SelectedRows.First();
             await NexusTable.CancelEditRow(data);
+        }
+
+        public void ShowInfoNexus()
+        {
+            if (NotificationService != null)
+                NotificationService.ShowInfo("I'm an INFO message", "Уведомление");
+        }
+        public void ShowSuccessNexus()
+        {
+            if (NotificationService != null)
+                NotificationService.ShowSuccess("I'm an SUCCESS message", "Успех");
+        }
+        public void ShowWarningNexus()
+        {
+            if (NotificationService != null)
+                NotificationService.ShowWarning("I'm an WARNING message", "Предупреждение");
+        }
+        public void ShowErrorNexus()
+        {
+            if (NotificationService != null)
+                NotificationService.ShowError("I'm an ERROR message", "Ошибка");
+        }
+
+        public async Task ShowDialog()
+        {
+            if (DialogService != null)
+            {
+                var settings = new NexusDialogSetting("Подтверждение", "Вы уверены, что хотите выполнить это действие?", "Отмена", "Подтвердить");
+                var result = await DialogService.Show(settings);
+                if (result?.Canceled == false)
+                {
+                    // Действие при подтверждении
+                    Console.WriteLine("Пользователь подтвердил действие");
+                }
+                else
+                {
+                    // Действие при отмене
+                    Console.WriteLine("Действие отменено");
+                }
+            }
         }
     }
 }
