+++ SibCCSPETest.Admin/Page/TopicPage.razor.cs
@@ -9,6 +9,10 @@ namespace SibCCSPETest.Admin.Page
     {
         [Inject]
         public IAPIService? ServiceAPI { get; set; }
+        [Inject]
+        public INexusNotificationService? NotificationService { get; set; }
+        [Inject]
+        public INexusDialogService? DialogService { get; set; }
 
         private NexusTableGrid<TopicDTO>? NexusTable;
         private NexusTableGridEditMode EditMode = NexusTableGridEditMode.Single;
@@ -86,6 +90,7 @@ namespace SibCCSPETest.Admin.Page
             {
                 NexusTable!.Data.Add(data);
                 await NexusTable.SelectRow(data);
+                NotificationService.ShowSuccess("Тема добавлена", "Успех");
             }
         }
 
@@ -99,6 +104,7 @@ namespace SibCCSPETest.Admin.Page
                     NexusTable.Data[index] = data;
                 await NexusTable.SelectRow(data);
                 await NexusTable.CancelEditRow(data);
+                NotificationService.ShowSuccess("Тема изменена", "Успех");
             }
         }
 
@@ -107,8 +113,14 @@ namespace SibCCSPETest.Admin.Page
             if (NexusTable != null && NexusTable.SelectedRows.Count != 0)
             {
                 var data = NexusTable.SelectedRows.First();
-                await ServiceAPI!.TopicService.DeleteTopic(data.Id);
-                NexusTable.RemoveRow(data);
+                var settings = new NexusDialogSetting("Удаление темы", $"Вы уверены, что хотите удалить \"{data.Title}\" тему?", "Отменить", "Удалить");
+                var result = await DialogService.Show(settings);
+                if (result?.Canceled == false)
+                {
+                    await ServiceAPI!.TopicService.DeleteTopic(data.Id);
+                    NexusTable.RemoveRow(data);
+                    NotificationService.ShowSuccess("Тема удалена", "Успех");
+                }
             }
         }
